backend/requirements.txt
fastapi
uvicorn[standard]
pillow
python-dotenv
tflite-runtime
numpy<2.0.0
google-generativeai
python-multipart
pydantic
motor
pymongo
opencv-python-headless
